Устраняем некоторые предупреждения компилятора: 1. Используем явное приведение целых типов с большей разрядности к меньшей. 2. В функции swscanf_s задаем размер буфера равный 1 при чтении одиночных символов "%c". 3. MpegSplitter: отключаем макрос MARKER перед его новым определением. 4. MPCVideoDec: отключаем предупреждение С4005 для макроса MAX_SLICES при включении файлов ffmpeg/libavcodec/. 5. DSUtil: устранена ошибка в функции ReadDisplay(), когда указатель приводился к типу UINT16.
git-svn-id: svn://svn.code.sf.net/p/mpcbe/code/trunk@3163 7426f7ed-705a-47fc-8cda-d59144bb0dce

+++ src/filters/source/FLACSource/FLACSource.cpp
@@ -406,7 +406,7 @@ HRESULT CFLACStream::GetMediaType(int iPosition, CMediaType* pmt)
 		wfe->nAvgBytesPerSec	= m_nAvgBytesPerSec;
 		wfe->nBlockAlign		= 1;
 		wfe->wBitsPerSample		= m_wBitsPerSample;
-		wfe->cbSize				= m_extradata.GetCount();
+		wfe->cbSize				= (WORD)m_extradata.GetCount();
 		memcpy(wfe + 1, m_extradata.GetData(), m_extradata.GetCount());
 	} else {
 		return VFW_S_NO_MORE_ITEMS;
